requirements.txt
Django==4.2.9
djangorestframework